Moving the i18n language setting to redux, which is persisted to local storage and removing react-use-cookie dependency
