+++ package.json
@@ -21,7 +21,6 @@
     "react-redux": "^8.0.2",
     "react-router-dom": "^6.3.0",
     "react-spinners": "^0.12.0",
-    "react-use-cookie": "^1.2.1",
     "redux-logger": "^3.0.6",
     "redux-persist": "^6.0.0",
     "semantic-ui-less": "^2.4.1",
